Move AutosaveErrorDialog to widgets directory
