+++ spyder/plugins/editor/utils/autosave.py
@@ -13,12 +13,11 @@ import os.path as osp
 
 # Third party imports
 from qtpy.QtCore import QTimer
-from qtpy.QtWidgets import (QCheckBox, QDialog, QDialogButtonBox, QLabel,
-                            QVBoxLayout)
 
 # Local imports
 from spyder.config.base import _, get_conf_path
 from spyder.config.main import CONF
+from spyder.plugins.editor.widgets.autosaveerror import AutosaveErrorDialog
 from spyder.plugins.editor.widgets.recover import RecoveryDialog
 
 
@@ -222,75 +221,10 @@ class AutosaveForStack(object):
         except EnvironmentError as error:
             action = (_('Error while autosaving {} to {}')
                       .format(finfo.filename, autosave_filename))
-            msgbox = AutosaveErrorMessageBox(action, error)
+            msgbox = AutosaveErrorDialog(action, error)
             msgbox.exec_if_enabled()
 
     def autosave_all(self):
         """Autosave all opened files."""
         for index in range(self.stack.get_stack_count()):
             self.autosave(index)
-
-
-class AutosaveErrorMessageBox(QDialog):
-    """
-    Dialog window notifying user of autosave-related errors.
-
-    The window also includes a check box which allows the user to hide any
-    future autosave-related errors.
-
-    Class attribute:
-        show_errors (bool): whether to show errors or not
-    """
-
-    show_errors = True
-
-    def __init__(self, action, error):
-        """
-        Constructor.
-
-        Args:
-            action (str): what Spyder was trying to do when error occured
-            error (Exception): the error that occured
-        """
-        logger.error(action, exc_info=error)
-
-        QDialog.__init__(self)
-        self.setWindowTitle(_('Autosave error'))
-        self.setModal(True)
-
-        layout = QVBoxLayout()
-
-        header = _('Error message:')
-        txt = '<br>{}<br><br>{}<br>{!s}'.format(action, header, error)
-        layout.addWidget(QLabel(txt))
-        layout.addSpacing(15)
-
-        txt = _("Hide all future autosave-related errors during this session")
-        self.dismiss_box = QCheckBox(txt)
-        layout.addWidget(self.dismiss_box)
-        layout.addSpacing(15)
-
-        button_box = QDialogButtonBox(QDialogButtonBox.Ok)
-        button_box.accepted.connect(self.accept)
-        layout.addWidget(button_box)
-
-        self.setLayout(layout)
-
-    def exec_if_enabled(self):
-        """
-        Execute dialog box unless disabled by the user.
-
-        The dialog box is disabled once the user clicks the 'Hide all future
-        errors' check box on one dialog box.
-        """
-        if AutosaveErrorMessageBox.show_errors:
-            return self.exec_()
-
-    def accept(self):
-        """
-        Update `show_errors` and hide dialog box.
-
-        Overrides method of `QDialogBox`.
-        """
-        AutosaveErrorMessageBox.show_errors = not self.dismiss_box.isChecked()
-        return QDialog.accept(self)
